#23 - moved mocking alongside service
